fix: se repararon errores de logica y nomenclatura menores en los msvc´s review, shipping y product

+++ msvc_product/src/main/java/com/GameHub/models/dtos/ProductRequestDTO.java
@@ -22,7 +22,7 @@ public class ProductRequestDTO {
 
     @NotNull(message = "El precio es obligatorio.")
     @Positive(message = "El precio debe ser un número mayor que cero.")
-    private BigDecimal precio;
+    private Long precio;
 
     @NotNull(message = "El ID de la categoria es obligatorio.")
     private Long categoriaId;
